AbstractGetCodeOwnersForPath: Do filtering only once at the end
Currently we filter code owners in multiple places at the moment when we
collect them. Instead, now we do the filtering only once at the end
after all code owners have been collected.

Earlier we aborted the collection of code owners when the limit had been
reached, hence the filtering had to be done immediately before the
collection of code owners was aborted as otherwise we would have
returned fewer suggestions than the limit if results were filtered out
later. However nowadays we collect all code owners and do the limiting
only at the end (with a few exeptions that we remove now), hence now we
can also do the filtering at the end.

Doing the filtering at the end removes duplicate calls of the filter
method and also allows us to implement additional filtering based on
annotations in a follow-up change (we first need to collect annotations
from all relevant code owner config files before we can use them
for filtering).

When filling up code owners with random users (when the code ownership
is assigned to all users aka '*') we:

* make sure to fill up the limit
* prefer explicitly mentioned code owners with the same score over
  random code owners

With the old logic this was achieved by:
1. Get 2 times more users than needed to fill up the limit
2. Do the filtering (e.g. filter out change owner)
3. Drop all random users that are over the limit

Since with 3. we dropped random code owners that were over the limit, we
ensured that explicitly mentioned code owners with the same score always
appeared in the suggestion.

Keeping this functionality intact required a few extra changes.

With this change we do the filtering only at the very end, hence step 2.
is dropped. This means we also have to drop step 3. as otherwise it can
happen that one of the randomly selected code owners is filtered out
later (e.g. because it was the change owner) and then there are not
enough suggestions to fill the limit. If step 3. is dropped we can now
have more code owners than the limit and some of them are dropped later.
Which of them are dropped is random since they all have the same
distance score. This is an issue if there are also explicitly mentioned
code owners with the same distance score, as they can be dropped too.
This should not happen, since we prefer explicitly mentioned code owners
over randomly selected code owners. To ensure that the explicitly
mentioned code owners are not dropped in this case, we add a new score
for them, so that they have a higher total score than randomly selected
code owners on the same level, and hence are no longer dropped.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib083bd2c1010d000ffa5cdb4f7edc93443d1ef62