AbstractGetCodeOwnersForPath: Drop unneeded if conditions for debug logs
We only populate the debug logs in the JSON if the debug flag was set to
true. This means it's okay to collect debug logs even if the debug flag
is false, as in this case they are not returned to the caller. This
simplifies the code a little bit.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Icfbfda02b48e0aee6c031bcb2c6ae8834d9a4ab5