Suggest code owners with NEVER_SUGGEST if the result would be empty
This is a requirement from the Android team (for which we added the
NEVER_SUGGEST annotation in the first place). If the suggestion result
would be empty if code owners that are annotated with NEVER_SUGGEST are
ignored, the NEVER_SUGGEST annotation should have no effect which means
that in this case the code owners that are annotated with NEVER_SUGGEST
should be suggested. This is a bit confusing since the annotation name
suggests that these code owners will never be suggested, but now there
is one exception for this. To make this less bad we may consider
renaming the annotation in a follow-up change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id71314cbd0216aaab59f7b356a23151f9b12f040