Rename NEVER_SUGGEST annotation to LAST_RESORT_SUGGESTION
With change Id71314cbd the behavior of the NEVER_SUGGEST annotation was
changed, so that code owners the are annotated with NEVER_SUGGEST are
suggested if otherwise the suggestion result would be empty. This is a
bit confusing since this behavior doesn't match the name of the
NEVER_SUGGEST annotation which implies that code owners with this
annotation would never be suggested. To remove this confusion the
NEVER_SUGGEST annotation is renamed to LAST_RESORT_SUGGESTION which
describes the behavior much better. Code owners with this annotation are
suggested as a last resort if no other code owners are available.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I74726784ee7138f40aeb9d783e5ee52c917f2ce7