Test skipping code owner config validation when creating merge via API
If a change for a merge commit is created via the Create Change REST
endpoint the code owner config file validation validates any OWNERS file
that is touched in the merge commit. This validation fails if such an
OWNERS file contains conflicts (conflicts are possible if
mergeInput.allowConflicts = true was set), since the conflict markers in
the file will be detected as invalid content.

Change ... in Gerrit core allows to set validation options in the change
input that are forwarded to the CommitValidationListeners. This makes it
possible to skip the code owner config validation when creating a change
via the Create Change REST endpoint, by specifying
code-owners~skip-validation=true as validation option.

This change adds a test to verify that skipping the code owner config
validation by setting code-owners~skip-validation=true in change input
works.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib9eb5182cf0aabfc07b61a72ba95edf259222de3