Add a manually maintained index documentation page
This will be used at Google to generate the index.html page. This is
needed because at Google we do not have plugin jar files, but plugin
documentation files are directly loaded from the Gerrit jar where we
cannot list files of a folder. Hence we cannot detect dynamically which
documentation files exist so that they can be listed in the index.html
page. Instead we just include toc.md if it exists.

Initially the references in toc.md are sorted by file name, the same way
as the references would show up in an automatically generated
index.html. Since we maintain toc.md manually we can find a better
sorting and logical grouping, but this we will do in follow-up changes.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5d790cf40d9fe54400de06bfb3961b23dcad5e41