CodeOwnersSubmitRule: Treat invalid plugin configuration as RULE_ERROR
In case of a failure submit rules can throw an exception or return a
RULE_ERROR.

An exception should be thrown in case of server errors that should count
against Gerrit’s SLO’s and that need to be investigated by the Gerrit
team.

A RULE_ERROR should be returned in case a failure is caused by bad
configuration that is under the control of host and project owners. In
this case it’s the responsibility of the host/project owner to fix the
configuration and the error should not count against Gerrit's SLO’s.

Since the code-owners plugin configuration is maintained by the
host/project owners, any failure that is caused by a bad configuration
should be treated as RULE_ERROR, but so far we threw an exception.

Fixing this allows us to run the code-owners tests that verify the
behaviour when the plugin configuration is invalid against the fake
index. Before this change these tests failed when running against the
fake index because the fake index is more strict when exceptions occur
during the indexing. With the fake index exceptions are propagated,
whereas with other index implementations they are swallowed (which means
that the exception is caught, but the change is not indexed). For the
code-owners plugin this meant that creating changes was possible when
the code-owners plugin configuration was invalid, but these changes
didn't get indexed. With this fix, they are indexed now and only the
code-owners submit requirement in the index says RULE_ERROR.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1ec13548b42f541cc40fcc651e997d374b08ee02