Test the extended change messages on post review are included in emails
If an email notification is sent on post review it includes the change
message that is being added to the change. If a code owner approval or a
code owner override was applied the code-owners plugin extends this
change message. Add a test to verify that change message extensions are
present in the email notification.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I70e2dd61ff1e0e35d463db671b21b9ea1236f590