Suggest owners that are reviewers even if annotated with LAST_RESORT_SUGGESTION
So far code owners that are annotated with LAST_RESORT_SUGGESTION were
only suggested if the suggestion result would be empty otherwise. This
meant that these code owners were also not suggested after they have
been manually added as a reviewer. Due to this the UI couldn't show
files that were owned by these users with a check mark (= covered by a
reviewer) as the UI only considers code owners that are part of the
suggestion. To fix this issue, code owners that are annotated with
LAST_RESORT_SUGGESTION are now also included in the suggestion result if
they are already a reviewer. This means if a code owner is already a
reviewer the LAST_RESORT_SUGGESTION annotation is ignored.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6b6a0a7d776e166d655076019bbc3596a713517f