Add disclaimer to find-owners docs that examples use FIND_OWNERS_GLOB's
Since change Ida22f2ae8 the path expression syntax can be freely
configured. All examples of OWNERS files in the find-owners
documentation assume that FIND_OWNERS_GLOB's are used as path
expressions. Add a note about this to avoid confusion for user of
projects that use the find-owners backend with a different path
expression syntax.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id00335ca07d74bc9b6b31a669f9160ec882e5da2