ChangedFiles: Check experiment flag only when current user is available
Checking whether the experiment flag for using the diff cache is set
requires a current user, at least with the Google implementation of
experiment flags. If a current user is not available a warning is logged
and the experiment flag is considered as not set. Since ChangedFiles is
called from a background thread regularly (e.g. when indexing a change)
we see this warning frequently. Fix this by checking for the experiment
flag only when a current user is available, otherwise return false to
not use the diff cache in this case.

Alternatively it was considered to make a current user available by
opening a ManualRequestContext, but we don't have a suitable user
available that should be provided as current user and providing the
internal user would always evaluate to false, since experiment flags are
never set for the internal user. This means we better directly return
false in this case.

Bug: Google b/191961988
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia0359f5e8ea661df734b53739772e4b2282e2bcc