Add a has:enabled_code-owners and has:approval_code-owners operands
Adding two operands to the 'has' predicate:
  * has:enabled_code-owners
  * has:approval_code-owners

The CodeOwnerApprovalPredicate wraps the existing CodeOwnerSubmitRule
logic. The predicate returns true only if the submit rule returns a
record with status = "OK". We will need this predicate to create query
expressions that depend on code owner approvals in submit requirements.

The two new operators work only for submit requirement expressions and
are not supported in search queries.

Bug: Google b/200665812
Change-Id: I391a0d7fd43528c87ff9dfc7ca981c7dc4d05407