Do not log invalid override configuration as warning
If overrides are configured incorrectly this is not a problem for Gerrit
since in this case overrides are simply disabled, but it doesn't cause
errors.

Logging a warning in this case adds too much noise in logs, hence reduce
to info logging and do not log the stacktrace which is never needed.

This warning would only be interesting to the project owners, but they
don't have access to the Gerrit logs and hence are not able to see this
warning. To see this information in traces, logging on info level is
sufficient.

Bug: Google b/205942673
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I737fef931377a285134ac20a09e117e768078b1c