Adapt to API change in Gerrit core and handle case where branch rev not present
Change I60ced9e37 in Gerrit core changed

  BranchResource#getRevision()

from returning String to returning Optional<String> to force callers to
handle the case where the revision is not present. As explained in the
commit message of change I60ced9e37 the revision is not present if the
branch identifier in a REST call was a symbolic ref that points to an
unborn branch.

Handle and test this case for all branch REST endpoints.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibf85ff0a6511b47806f4bd2fbf496b91a1622ea9