CodeOwnerSubmitRule: Export the rule name
With change I3f08f189d we added a ruleName field to the SubmitRecord
entity, this is to ease the conversion of submit records to new submit
requirements and give them a meaningful name.

Exporting a meaningful rule name in this change so that it could be ued
By Gerrit core.

Change-Id: I6ef833f33fe1e1d4a1f3086349d701fe1ce514f8