Fix NPE if OWNERS file imports '/' instead of '/OWNERS'
An import in an OWNERS file must specify the file name of the imported
OWNERS file. This means to import the OWNERS file from the root
directory '/OWNERS' must be imported. Trying to attempt importing the
root folder '/' failed with a NPE because the file name of the path is
null in this case. Fix the NPE and return a proper error message in this
case.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib92dc40e007ce7a9693896cf394d57adc844db07