Make log less verbose if change message extention fails due to config error
CodeOwnersOnAddReviewer extends the change message when a code owner is
added as a reviewer. It lists the owned paths in the message. If any
relevant OWNERS file is invalid and cannot be parsed, we cannot compute
the owned files, and hence the message cannot be extended in this case.
Adding the reviewer succeeds but only the change message that is posted
for it doesn't contain any owned files. This works as intended, but we
log a verbose error (with stacktrace) in this case that is not needed.
Fix the exeption handling so that we log a warning only if the exception
is caused by a configuration error. Otherwise still log the verbose
error so that it can be investigated when it happens.

OnCodeOwnerApproval extends the change message when a code owner
approval is applied. Similarly to CodeOwnersOnAddReviewer we cannot
extend the message if relevant OWNERS files are invalid. Here too the
exception was ignored and logged with stacktrace. Catching and logging
this exception was done generically by PluginContext. Now we are
handling the exception on our own, so that also here we can log a
warning only when the exception is caused by a configuration error.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I0d1b3cf747643fd7bef512e6dce0fbd8bb940da9