Test skipping code owner config validation when creating cherry pick via API
If a change / patch set is created via the Cherry Pick Revision REST
endpoint the code owner config file validation validates any OWNERS file
that is touched in the commit. This validation fails if such an OWNERS
file contains conflicts (conflicts are possible if
cherryPickInput.allowConflicts = true was set), since the conflict
markers in the file will be detected as invalid content.

Change I66dc4cfda in Gerrit core allows to set validation options in the
cherry pick input that are forwarded to the CommitValidationListeners.
This makes it possible to skip the code owner config validation when
creating a change via the Cherry Pick Revision REST endpoint, by
specifying code-owners~skip-validation=true as validation option.

This change adds a test to verify that skipping the code owner config
validation by setting code-owners~skip-validation=true in the cherry
pick input works.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1f6b2f8a3ac1c821510f7ec8e294f01d72468abf